% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccscore.R
\name{alphazeroccscore}
\alias{alphazeroccscore}
\title{Hscore alpha = 0}
\usage{
alphazeroccscore(mat)
}
\arguments{
\item{mat}{a matrix}
}
\value{
The hscore of \code{mat}
}
\description{
Calculates Hscore with alpha equal to 0
}
\examples{
mat <- replicate(5, rnorm(5))
alphazeroccscore(mat)
}
