% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{bic_map}
\alias{bic_map}
\title{Biclusters Map}
\usage{
bic_map(mat, bic_list)
}
\arguments{
\item{mat}{a matrix}

\item{bic_list}{a list of biclusters from \code{hc}}
}
\value{

}
\description{
Creates a two-dimensional plot where every dot is a bicluster
  whose length of the sub-interval is represented by the x-axis, numbers of
  functions by the y-axis and Hscore by the dimension of the dot itself, i.e.
  small dot refers to low Hscore while big dot is for results having greater
  scores.
}
\examples{


}
