% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccscore.R, R/hello.R
\name{ccscore}
\alias{ccscore}
\title{Hscore}
\usage{
ccscore(mat)

ccscore(mat)
}
\arguments{
\item{mat}{a matrix}

\item{mat}{a matrix}
}
\value{
The hscore of \code{mat}

The hscore of \code{mat}
}
\description{
Calculates Hscore with both alpha and beta different to 0

Calculate Hscore with both alpha and beta both different to 0
}
\examples{
mat <- replicate(5, rnorm(5))
ccscore(mat)
mat <- replicate(5, rnorm(5))
ccscore(mat)
}
