% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccscore.R
\name{d_ccscore}
\alias{d_ccscore}
\title{Two functions Hscore}
\usage{
d_ccscore(x, y)
}
\arguments{
\item{x}{a function}

\item{y}{a function}
}
\value{
The hscore of a matrix having \code{x} and \code{y} as rows
}
\description{
Calculates the Hscore of the matrix composed by function \code{x}
  and function \code{y} as rows
}
\examples{
x <- runif(5, min=0, max=100)
y <- runif(5, min=0, max=100)
d_ccscore(x,y)
}
