% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvesting.R
\name{deletedanger}
\alias{deletedanger}
\title{Delete Inversions}
\usage{
deletedanger(hc)
}
\arguments{
\item{hc}{a hierarchy resulting from DIANA}
}
\value{
A vector with the new hscore-based height to avoid having branches
  inversions in the dendogram
}
\description{
Identifies and solves branches Inversions
}
\examples{

}
