% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvesting.R
\name{delta_cutter}
\alias{delta_cutter}
\title{Delta-Threshold Cutter}
\usage{
delta_cutter(mat, hc, delta)
}
\arguments{
\item{mat}{a matrix}

\item{hc}{a hierarchy resulting from DIANA built from \code{mat}}

\item{delta}{a scalar value indicating the delta-threshold}
}
\value{
A list containing all the candidate biclusters whose Hscore is lesser
  than \code{delta}. Every single element of the list is a candidate
  bicluster and the elements belonging to it are reprented using the
  row index/label of the matrix \code{mat}.
}
\description{
Cuts a Hscore dendrogram according to the threshold \code{delta}
  finding biclusters whose hscore is lesser than \code{delta}.
}
