% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dendro_df}
\alias{dendro_df}
\title{Dendrogram Dataframe Creator}
\usage{
dendro_df(hc)
}
\arguments{
\item{hc}{a hierarchy resulting from DIANA}
}
\value{
A hierarchy dataframe suitable for ggplot2 visualization
}
\description{
Transforms the output of a hierarchical algorithm \code{hc} in a dataframe
  suitable for ggplot2 visualization
}
\examples{
mat <- replicate(5, rnorm(5))
hc  <- funBI_diana(mat)
dendro_df(hc)
}
