% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotting_flowering.R
\name{funBI_diana}
\alias{funBI_diana}
\title{Diana Tree}
\usage{
funBI_diana(mat)
}
\arguments{
\item{mat}{a matrix}
}
\value{
A DIANA hierarchy, i.e. a DIANA list with hscore height
}
\description{
Performs hscore based DIANA (DIvisive ANAlysis) algorithm
}
\examples{
mat <- replicate(20, rnorm(15))
funBI_diana(mat)
}
