% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvesting.R
\name{funBI_harvesting}
\alias{funBI_harvesting}
\title{funBI Harvesting}
\usage{
funBI_harvesting(list_hc, delta, perc, gap, mat, Kmax, B)
}
\arguments{
\item{list_hc}{a list of hierarchy resulting from DIANA built from \code{mat}}

\item{delta}{a scalar value indicating the delta-threshold in order to
perform a delta-bicluster. In this case \code{perc} and \code{gap} should
be missing.}

\item{perc}{a scalar value indicated a percentage of the total score in order
to perform a delta-bicluster cut  using as delta-threshold the
percentage \code{perc} of the total hscore of a particular dendrogram. In
this case \code{delta} and \code{gap} should be missing.}

\item{gap}{TRUE/FALSE indicating if gap statistics should be used as
Harvesting strategy.In this case \code{delta} and \code{perc} should be
missing.}

\item{mat}{a matrix}

\item{Kmax}{the maximum number of clusters to consider when performing gap
statistics (\code{gap} = TRUE), must be at least two.}

\item{B}{integer value, number of Monte Carlo (“bootstrap”) samples to
consider when performing gap statistics (\code{gap} = TRUE).}
}
\value{
\code{biclist}, a list containing all the candidate biclusters
  according to the used Harvesting strategy. Every element of the list is a
  bicluster. Every bicluster is characterized by the following components:
  \enumerate{ \item{a vector containing the elements belonging to the
  bicluster reprented using the row index/label of the matrix \code{mat}}
  \item{a vector containing the boundaries of the sub-interval
  considered represented usign the colum index/label of the matrix
  \code{mat}} \item{the Hscore of the bicluster} \item{The index
  of \code{list_hc}, representing which dendogram the bicluster was cut} }
}
\description{
Perform Harvesting to a list of DIANA hierarcy obetained using
  the \code{funBI_lot_and_flower} function. The available harvesting
  strategies are three: a \code{delta}-threshold strategy, a dynamic strategy
  with a threshold obtained as a percentage of the total hscore of a
  considered hierarchical dendrogram, and an automatic strategy using
  \code{gap} statistics.
}
\examples{

}
