% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotting_flowering.R
\name{funBI_lot_and_flower}
\alias{funBI_lot_and_flower}
\title{Lotting and Flowering}
\usage{
funBI_lot_and_flower(mat, int)
}
\arguments{
\item{mat}{a matrix}

\item{int}{the minimum length of sub-intervarls}
}
\value{
A list of DIANA hierarchy, one for each sub-intervals, with the
  reference of starting and ending of the corresponding intervals
}
\description{
Performs the Lotting and Flowering steps by dividing the whole
  domain in sub-intervals of length greater than \code{int} and by applying
  DIANA in every sub-interval
}
\examples{
mat <- replicate(5, rnorm(5))
int <- 2
funBI_lot_and_flower(mat,int)
}
