% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotting_flowering.R
\name{funBI_lotting}
\alias{funBI_lotting}
\title{Lotting}
\usage{
funBI_lotting(mat, int)
}
\arguments{
\item{mat}{a matrix}

\item{int}{the minimum length of sub-intervarls}
}
\value{
A list containing all the possible starts and lengths of the sub-intervals
}
\description{
Divides the domain in sub-intervals of lenght greater than
  \code{int}. The smallest sub-intervals have length equal to \code{int} while the
  others are created by shifting by \code{int} along the continuous dimension and by
  enlarging them by multiples of \code{int} in order to cover the whole T
}
\examples{
mat <- replicate(5, rnorm(5))
int <- 2
funBI_lotting(mat, int)
}
