% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasting.R
\name{funBI_tasting}
\alias{funBI_tasting}
\title{Delta-Threshold Cutter}
\usage{
funBI_tasting(biclist)
}
\arguments{
\item{biclist}{a list of candidate biclusters, resulting from
\code{funBI_harvesting}}
}
\value{
A reduced version of \code{biclist} containing all the candidates
  bicluster which are not nested into other candidates biclusters.
}
\description{
Cuts a Hscore dendrogram according to the threshold \code{delta}
  finding biclusters whose hscore is lesser than \code{delta}.
}
