% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvesting.R
\name{gap_clus}
\alias{gap_clus}
\title{Gap Statistics}
\usage{
gap_clus(mat, hc, K.max, B, d.power = 1, spaceH0 = c("scaledPCA",
  "original"))
}
\arguments{
\item{mat}{a matrix}

\item{hc}{a hierarchy resulting from DIANA}

\item{K.max}{the maximum number of clusters to consider, must be at least
two.}

\item{B}{integer value, number of Monte Carlo (“bootstrap”) samples}
}
\value{
An object of S3 class "clusGap", basically a list with components:
  \describe{ \item{Tab}{a matrix with K.max rows and 4 columns, named "logW",
  "E.logW", "gap", and "SE.sim"} \item{call}{the function call}
  \item{spaceH0}{the \code{spaceH0} argument} \item{n}{number of observation}
 \item{B}{input \code{B}}}
}
\description{
calculates a goodness of clustering measure, the \code{gap} statistic.
}
\examples{

}
