% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvesting.R
\name{hcut}
\alias{hcut}
\title{Cutter}
\usage{
hcut(new_height, mergelist, delta, mat)
}
\arguments{
\item{new_height}{a vector with the hscore-based height (modified or not) of
the dendogram built from \code{delta}}

\item{mergelist}{a list of all the possible biclusters one can get from \code{hc}}

\item{delta}{a scalar value indicating the delta-threshold}

\item{mat}{a matrix}
}
\value{
A list containing all the candidate biclusters whose Hscore is lesser
  than \code{delta}. Every single element of the list is a candidate
  bicluster and the elements belonging to it are reprented using the
  row index/label of the matrix \code{mat}.
}
\description{
Cuts a Hscore dendrogram according to the threshold \code{delta}
  finding biclusters whose hscore is lesser than \code{delta}.
}
\examples{

}
