% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{hscoredendroplot}
\alias{hscoredendroplot}
\title{ggplot2 Hscore Dendrogram Plot}
\usage{
hscoredendroplot(hc)
}
\arguments{
\item{hc}{a hierarchy resulting from DIANA}
}
\value{
ggplot object
}
\description{
Creates from the output of a hierarchical algorithm \code{hc} a
  dataframe suitable for a ggplot2 visualization using dendro_df. Then it
  plots a ggplot2 dendrogram having hscore as height
}
\examples{

}
