% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotting_flowering.R
\name{score_merging}
\alias{score_merging}
\title{Score Merger}
\usage{
score_merging(mat, hc)
}
\arguments{
\item{mat}{a matrix}

\item{hc}{a hierarchy resulting from DIANA}
}
\value{
A list containing the hscores of every single DIANA merging steps and
  the resulting groups of those steps
}
\description{
Gets the height of every single DIANA merging step
}
